Move `fj::Group` to dedicated module
